requirements.txt
# Pinned from the configured .venv used for this repository.
# Install with:
#   python -m pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu124
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gdown==5.2.1
h5py==3.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
soupsieve==2.8.3
sympy==1.13.1
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1